+++ packages/atomic/src/components/common/pager/pager-navigation.tsx
@@ -11,7 +11,7 @@ export const PagerNavigation: FunctionalComponent<PagerNavigationProps> = (
 ) => {
   return (
     <nav aria-label={props.i18n.t('pagination')}>
-      <div part="buttons" class="flex flex-wrap gap-2">
+      <div part="buttons" role="toolbar" class="flex flex-wrap gap-2">
         {...children}
       </div>
     </nav>