+++ packages/headless-react/src/ssr-commerce/providers.tsx
@@ -1,6 +1,7 @@
 'use client';
 
 import {
+  HydrateStaticStateOptions,
   InferHydratedState,
   InferStaticState,
   NavigatorContext,
@@ -52,6 +53,15 @@ export function buildProviderWithDefinition(looseDefinition: LooseDefinition) {
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       const hydrateControllers: Record<string, any> = {};
 
+      if ('parameterManager' in controllers) {
+        hydrateControllers.parameterManager = {
+          initialState: {
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            parameters: (controllers as any).parameterManager.state.parameters,
+          },
+        };
+      }
+
       if ('cart' in controllers) {
         hydrateControllers.cart = {
           // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -71,7 +81,7 @@ export function buildProviderWithDefinition(looseDefinition: LooseDefinition) {
             ...controllers,
             ...hydrateControllers,
           },
-        })
+        } as HydrateStaticStateOptions<{type: string}>)
         .then(({engine, controllers}) => {
           setHydratedState({engine, controllers});
         });