+++ packages/headless/esbuild.mjs
@@ -295,6 +295,7 @@ const nodeEsm = Object.entries(useCaseEntries).map((entry) => {
     {
       entryPoints: [entryPoint],
       outfile,
+      sourcemap: true,
       format: 'esm',
       external: ['pino'],
       mainFields: ['module', 'main'],