+++ packages/headless/src/app/commerce-ssr-engine/types/common.ts
@@ -5,23 +5,23 @@ import type {
   HasKey,
   InferControllerStaticStateMapFromControllers,
   InferControllerStaticStateFromController,
-  InferControllerPropsMapFromDefinitions,
   ControllerStaticStateMap,
   EngineDefinitionBuildResult,
-  EngineDefinitionControllersPropsOption,
   HydratedState,
   OptionsTuple,
+  ControllersPropsMap,
+  HasKeys,
+  HasRequiredKeys,
+  HasOptionalKeys,
 } from '../../ssr-engine/types/common.js';
 import {SSRCommerceEngine} from '../factories/build-factory.js';
 
 export type {
   EngineDefinitionBuildResult,
-  EngineDefinitionControllersPropsOption,
   HydratedState,
   OptionsTuple,
   InferControllerStaticStateFromController,
   InferControllerStaticStateMapFromControllers,
-  InferControllerPropsMapFromDefinitions,
 };
 
 export enum SolutionType {
@@ -30,6 +30,21 @@ export enum SolutionType {
   standalone = 'standalone',
   recommendation = 'recommendation',
 }
+
+const recommendationOptionKey = 'recommendation-internal-options';
+export const recommendationInternalOptionKey = Symbol.for(
+  recommendationOptionKey
+);
+
+export type RecommendationControllerSettings = {
+  /**
+   * Toggle to enable or disable the recommendation controller.
+   * When set to `true`, the controller will perform a recommendation request server-side.
+   *
+   * @default false
+   */
+  enabled?: boolean;
+};
 
 export interface ControllerDefinitionWithoutProps<
   TController extends Controller,
@@ -71,9 +86,19 @@ export interface EngineStaticState<
   controllers: TControllers;
 }
 
+export interface SolutionTypeAvailability {
+  [SolutionType.search]?: boolean;
+  [SolutionType.listing]?: boolean;
+  [SolutionType.standalone]?: boolean;
+  [SolutionType.recommendation]?: boolean;
+}
+
 export type ControllerDefinition<TController extends Controller> =
-  | ControllerDefinitionWithoutProps<TController>
-  | ControllerDefinitionWithProps<TController, unknown>;
+  SolutionTypeAvailability &
+    (
+      | ControllerDefinitionWithoutProps<TController>
+      | ControllerDefinitionWithProps<TController, unknown>
+    );
 
 export interface ControllerDefinitionsMap<TController extends Controller> {
   [customName: string]: ControllerDefinition<TController>;
@@ -83,10 +108,27 @@ export type InferControllerPropsFromDefinition<
   TController extends ControllerDefinition<Controller>,
 > =
   TController extends ControllerDefinitionWithProps<Controller, infer Props>
-    ? Props
+    ? HasKey<TController, typeof recommendationInternalOptionKey> extends never
+      ? Props
+      : Props & RecommendationControllerSettings
     : TController extends ControllerDefinitionWithoutProps<Controller>
-      ? {}
+      ? HasKey<
+          TController,
+          typeof recommendationInternalOptionKey
+        > extends never
+        ? {}
+        : RecommendationControllerSettings
       : unknown;
+
+export type InferControllerPropsMapFromDefinitions<
+  TControllers extends ControllerDefinitionsMap<Controller>,
+> = {
+  [K in keyof TControllers as HasKeys<
+    InferControllerPropsFromDefinition<TControllers[K]>
+  > extends false
+    ? never
+    : K]: InferControllerPropsFromDefinition<TControllers[K]>;
+};
 
 export type InferControllerFromDefinition<
   TDefinition extends ControllerDefinition<Controller>,
@@ -121,6 +163,216 @@ export type InferControllerStaticStateMapFromDefinitionsWithSolutionType<
   >;
 };
 
+/**
+ * This type defines the required and optional controller props for the engine definition.
+ */
+export type EngineDefinitionControllersPropsOption<
+  TControllers extends ControllerDefinitionsMap<Controller>,
+  TControllersPropsMap extends ControllersPropsMap,
+  TSolutionType extends SolutionType,
+> = OptionalEngineDefinitionControllersPropsOption<
+  TControllers,
+  TControllersPropsMap,
+  TSolutionType
+> &
+  RequiredEngineDefinitionControllersPropsOption<
+    TControllers,
+    TControllersPropsMap,
+    TSolutionType
+  >;
+
+/**
+ * Represents an optional engine definition for controller properties.
+ *
+ * This type is used to define a map of optional controller properties based on the provided
+ * controller definitions, controller properties map, and solution type.
+ *
+ * @template TControllers - A map of controller definitions.
+ * @template TControllersPropsMap - A map of controller properties.
+ * @template TSolutionType - The type of solution.
+ *
+ * The type iterates over the controller keys (defined in the engine definition) and includes only those keys where:
+ * 1. The controller can be used for a specific solution type (HasKey<TControllers[K], TSolutionType> is not 'never').
+ * 2. The controller properties have optional options (HasOptionalKeys<ConditionalControllerProps<...>> is true).
+ *
+ * @example
+ * Given the following controller definitions:
+ * ```
+ * const {recommendationEngineDefinition} = defineCommerceEngine({
+ *  controllers: {
+ *    popularViewed: defineRecommendations({
+ *      options: {slotId: 'slot-id'}
+ *    })
+ *  }
+ * });
+ * ```
+ *
+ * The following code will not throw an error since the 'popularViewed' controller props are optional
+ * ```
+ * recommendationEngineDefinition.fetchStaticState({
+ *  controllers: {
+ *    popularViewed: {enabled: true, productId: 'some-product-id'} // This is optional
+ *  }
+ * })
+ * ```
+ *
+ * The following code (with no arguments) is also valid because the 'popularViewed' controller props are optional, and there are no other required controller props in the engine definition.
+ * ```
+ * recommendationEngineDefinition.fetchStaticState()
+ * ```
+ */
+export type OptionalEngineDefinitionControllersPropsOption<
+  TControllers extends ControllerDefinitionsMap<Controller>,
+  TControllersPropsMap extends ControllersPropsMap,
+  TSolutionType extends SolutionType,
+> = {
+  [K in keyof TControllers as HasKey<
+    TControllers[K],
+    TSolutionType
+  > extends never
+    ? never
+    : HasOptionalKeys<
+          ConditionalControllerProps<
+            TControllers,
+            TControllersPropsMap,
+            TSolutionType,
+            K
+          >
+        > extends false
+      ? never
+      : 'controllers']?: ConditionalControllerProps<
+    TControllers,
+    TControllersPropsMap,
+    TSolutionType,
+    K
+  >;
+};
+
+/**
+ * Represents a type that defines the required controller properties for engine definition controllers.
+ *
+ * @template TControllers - A map of controller definitions.
+ * @template TControllersPropsMap - A map of controller properties.
+ * @template TSolutionType - The type of solution being used.
+ *
+ * The type iterates over the controller keys (defined in the engine definition) and includes only those keys where:
+ * 1. The controller can be used for a specific solution type (HasKey<TControllers[K], TSolutionType> is not 'never').
+ * 2. The controller properties have required options (HasRequiredKeys<ConditionalControllerProps<...>> is true).
+ *
+ * The resulting type maps the valid keys to their corresponding conditional controller properties.
+ *
+ * @example
+ * Given the following controller definitions:
+ * ```
+ * const {standaloneEngineDefinition} = defineCommerceEngine({
+ *  controllers: {cart: defineCart()},
+ * });
+ * ```
+ *
+ * The following code will not throw an error since the 'cart' controller props are required for the standalone engine definition:
+ * ```
+ * standaloneEngineDefinition.fetchStaticState({
+ *   controllers: {
+ *     cart: {initialState: {items: []}},
+ *   },
+ * })
+ * ```
+ *
+ */
+export type RequiredEngineDefinitionControllersPropsOption<
+  TControllers extends ControllerDefinitionsMap<Controller>,
+  TControllersPropsMap extends ControllersPropsMap,
+  TSolutionType extends SolutionType,
+> = {
+  [K in keyof TControllers as HasKey<
+    TControllers[K],
+    TSolutionType
+  > extends never
+    ? never
+    : HasRequiredKeys<
+          ConditionalControllerProps<
+            TControllers,
+            TControllersPropsMap,
+            TSolutionType,
+            K
+          >
+        > extends false
+      ? never
+      : 'controllers']: ConditionalControllerProps<
+    TControllers,
+    TControllersPropsMap,
+    TSolutionType,
+    K
+  >;
+};
+
+type IsRecommendationController<
+  TController extends ControllerDefinition<Controller>,
+> = HasKey<TController, typeof recommendationInternalOptionKey>;
+
+/**
+ * This type ensures that recommendation controller props are optional, while other controller props remain required.
+ *
+ * It works by checking if the controller definition includes the `recommendationInternalOptionKey`.
+ *
+ * - If the key is present, the controller props are made optional.
+ * - If the key is absent, the controller props are required.
+ *
+ * Example:
+ *
+ * ```typescript
+ * type ControllerProps = InferControllerPropsFromDefinition<{
+ *   recommendation: defineRecommendation({}),
+ *   cart: defineCart(),
+ * }>;
+ *
+ * // In this example:
+ * // - The `recommendation` controller props are optional.
+ * // - The `cart` controller props are required.
+ *
+ * const props: ControllerProps = {
+ *   cart: {initialState: {items: []}},
+ *   // recommendation props can be omitted
+ * };
+ * ```
+ */
+type RecommendationControllerProps<
+  TControllers extends ControllerDefinitionsMap<Controller>,
+  TControllersPropsMap extends ControllersPropsMap,
+  K extends keyof TControllers,
+> = {
+  [I in keyof TControllersPropsMap as I extends K
+    ? IsRecommendationController<TControllers[I]> extends never
+      ? never
+      : I
+    : never]?: TControllersPropsMap[I];
+} & {
+  [I in keyof TControllersPropsMap as I extends K
+    ? IsRecommendationController<TControllers[I]> extends never
+      ? I
+      : never
+    : never]: TControllersPropsMap[I];
+};
+
+type DefaultControllerProps<
+  TControllers extends ControllerDefinitionsMap<Controller>,
+  TControllersPropsMap extends ControllersPropsMap,
+  K extends keyof TControllers,
+> = {
+  [I in keyof TControllersPropsMap as I extends K
+    ? I
+    : never]: TControllersPropsMap[I];
+};
+
+type ConditionalControllerProps<
+  TControllers extends ControllerDefinitionsMap<Controller>,
+  TControllersPropsMap extends ControllersPropsMap,
+  TSolutionType extends SolutionType,
+  K extends keyof TControllers,
+> = TSolutionType extends SolutionType.recommendation
+  ? RecommendationControllerProps<TControllers, TControllersPropsMap, K>
+  : DefaultControllerProps<TControllers, TControllersPropsMap, K>;
+
 export interface ControllerDefinitionOption {
   /**
    * Whether the controller will be used in a product listing context.
@@ -132,6 +384,21 @@ export interface ControllerDefinitionOption {
    * @defaultValue true
    */
   search?: boolean;
+}
+
+interface NonRecommendationController {
+  /**
+   * @internal
+   */
+  [SolutionType.search]: true;
+  /**
+   * @internal
+   */
+  [SolutionType.listing]: true;
+  /**
+   * @internal
+   */
+  [SolutionType.standalone]: true;
 }
 
 interface UniversalController {
@@ -147,6 +414,10 @@ interface UniversalController {
    * @internal
    */
   [SolutionType.standalone]: true;
+  /**
+   * @internal
+   */
+  [SolutionType.recommendation]: true;
 }
 
 interface SearchOnlyController {
@@ -181,6 +452,17 @@ interface SearchAndListingController {
   [SolutionType.listing]: true;
 }
 
+interface ListingAndStandaloneController {
+  /**
+   * @internal
+   */
+  [SolutionType.listing]: true;
+  /**
+   * @internal
+   */
+  [SolutionType.standalone]: true;
+}
+
 export type SearchOnlyControllerDefinitionWithoutProps<
   TController extends Controller,
 > = ControllerDefinitionWithoutProps<TController> & SearchOnlyController;
@@ -189,6 +471,11 @@ export type SearchOnlyControllerDefinitionWithProps<
   TController extends Controller,
   TProps,
 > = ControllerDefinitionWithProps<TController, TProps> & SearchOnlyController;
+
+export type ListingAndStandaloneControllerWithoutProps<
+  TController extends Controller,
+> = ControllerDefinitionWithoutProps<TController> &
+  ListingAndStandaloneController;
 
 export type ListingOnlyControllerDefinitionWithoutProps<
   TController extends Controller,
@@ -212,9 +499,19 @@ export type RecommendationOnlyControllerDefinitionWithProps<
 
 export type NonRecommendationControllerDefinitionWithoutProps<
   TController extends Controller,
-> = ControllerDefinitionWithoutProps<TController> & UniversalController;
+> = ControllerDefinitionWithoutProps<TController> & NonRecommendationController;
 
 export type NonRecommendationControllerDefinitionWithProps<
+  TController extends Controller,
+  TProps,
+> = ControllerDefinitionWithProps<TController, TProps> &
+  NonRecommendationController;
+
+export type UniversalControllerDefinitionWithoutProps<
+  TController extends Controller,
+> = ControllerDefinitionWithoutProps<TController> & UniversalController;
+
+export type UniversalControllerDefinitionWithProps<
   TController extends Controller,
   TProps,
 > = ControllerDefinitionWithProps<TController, TProps> & UniversalController;