+++ packages/headless/src/controllers/commerce/core/parameter-manager/headless-core-parameter-manager.test.ts
@@ -1,6 +1,7 @@
 import {SchemaValidationError} from '@coveo/bueno';
 import {Parameters} from '../../../../features/commerce/parameters/parameters-actions.js';
 import {parametersDefinition} from '../../../../features/commerce/parameters/parameters-schema.js';
+import {parametersReducer as commerceParameters} from '../../../../features/commerce/parameters/parameters-slice.js';
 import {FacetValueState} from '../../../../features/facets/facet-api/value.js';
 import {buildRelevanceSortCriterion} from '../../../../features/sort/sort.js';
 import {buildMockCommerceState} from '../../../../test/mock-commerce-state.js';
@@ -20,7 +21,6 @@ describe('parameter manager', () => {
   const mockActiveParametersSelector = vi.fn();
   const mockRestoreActionCreator = vi.fn();
   const mockFetchProductsActionCreator = vi.fn();
-  const mockEnrichParameters = vi.fn();
 
   function initEngine(preloadedState = buildMockCommerceState()) {
     engine = buildMockCommerceEngine(preloadedState);
@@ -35,7 +35,6 @@ describe('parameter manager', () => {
       activeParametersSelector: mockActiveParametersSelector,
       restoreActionCreator: mockRestoreActionCreator,
       fetchProductsActionCreator: mockFetchProductsActionCreator,
-      enrichParameters: mockEnrichParameters,
       ...props,
     });
   }
@@ -55,60 +54,93 @@ describe('parameter manager', () => {
       expect(parameterManager).toBeTruthy();
     });
 
-    it('validates initial state against schema', () => {
-      expect(() =>
-        initParameterManager({
-          initialState: {parameters: {page: -1}},
-          parametersDefinition,
-        })
-      ).toThrow(SchemaValidationError);
+    it('by default, does not load the commerceParameters reducer', () => {
+      expect(engine.addReducers).not.toHaveBeenCalledWith({
+        commerceParameters,
+      });
     });
 
-    it('dispatches #restoreActionCreator with initial parameters', () => {
-      initParameterManager({initialState: {parameters: {page: 2}}});
-
-      expect(mockRestoreActionCreator).toHaveBeenCalledWith({
-        page: 2,
+    it('when excludeDefaultParameters is set to false, does not load the commerceParameters reducer', () => {
+      initParameterManager({excludeDefaultParameters: false});
+      expect(engine.addReducers).not.toHaveBeenCalledWith({
+        commerceParameters,
       });
+    });
+
+    it('when excludeDefaultParameters is set to true, loads the commerceParameters reducer', () => {
+      initParameterManager({excludeDefaultParameters: true});
+      expect(engine.addReducers).toHaveBeenCalledWith({
+        commerceParameters,
+      });
+    });
+
+    describe('when an initial state is provided', () => {
+      it('validates initial state against schema', () => {
+        expect(() =>
+          initParameterManager({
+            initialState: {parameters: {page: -1}},
+            parametersDefinition,
+          })
+        ).toThrow(SchemaValidationError);
+      });
+
+      it('dispatches #restoreActionCreator with initial parameters', () => {
+        initParameterManager({initialState: {parameters: {page: 2}}});
+
+        expect(mockRestoreActionCreator).toHaveBeenCalledWith({
+          page: 2,
+        });
+      });
+    });
+  });
+
+  describe('when an initial state is not provided', () => {
+    it('does not dispatch #restoreActionCreator', () => {
+      mockRestoreActionCreator.mockReset();
+      initParameterManager({initialState: undefined});
+
+      expect(mockRestoreActionCreator).not.toHaveBeenCalled();
     });
   });
 
   describe('#synchronize', () => {
-    describe('when new parameters are the same as the old ones', () => {
-      it('does not dispatch any action', () => {
-        mockRestoreActionCreator.mockReset();
+    beforeEach(() => {
+      mockRestoreActionCreator.mockReset();
+      mockFetchProductsActionCreator.mockReset();
+    });
+    it('when the new parameters are the same as the previous ones, does not dispatch any action', () => {
+      mockRestoreActionCreator.mockReset();
+      mockFetchProductsActionCreator.mockReset();
 
-        const parameters = {page: 2};
-        parameterManager.synchronize(parameters);
+      mockActiveParametersSelector.mockReturnValue({page: 2});
+      parameterManager.synchronize({page: 2});
 
-        expect(mockRestoreActionCreator).not.toHaveBeenCalled();
-      });
+      expect(mockRestoreActionCreator).not.toHaveBeenCalled();
+      expect(mockFetchProductsActionCreator).not.toHaveBeenCalled();
     });
 
-    describe('when there is a difference in parameters', () => {
-      beforeEach(() => {
-        mockActiveParametersSelector.mockReturnValue({page: 1});
-        mockEnrichParameters.mockImplementation((_, params) => params);
-      });
+    it('when there are no new and no active parameters, dispatches #restoreActionCreator and #fetchProductsActionCreator', () => {
+      mockActiveParametersSelector.mockReturnValue({});
+      parameterManager.synchronize({});
 
-      it('dispatches #restoreActionCreator', () => {
-        const parameters = {
-          page: 2,
-        };
-        parameterManager.synchronize(parameters);
-        expect(mockRestoreActionCreator).toHaveBeenCalledWith(parameters);
-      });
+      expect(mockRestoreActionCreator).toHaveBeenCalled();
+      expect(mockFetchProductsActionCreator).toHaveBeenCalled();
+    });
 
-      it('dispatches #fetchProductsActionCreator', () => {
-        parameterManager.synchronize({page: 2});
-        expect(mockFetchProductsActionCreator).toHaveBeenCalled();
-      });
+    it('when there the new parameters are different from the previous ones, dispatches #restoreActionCreator and #fetchProductsActionCreator', () => {
+      mockActiveParametersSelector.mockReturnValue({page: 2});
+      parameterManager.synchronize({perPage: 5});
+
+      expect(mockRestoreActionCreator).toHaveBeenCalled();
+      expect(mockFetchProductsActionCreator).toHaveBeenCalled();
     });
   });
 
   describe('#state', () => {
-    it('contains #parameters', () => {
-      const parameters = {
+    let parameters: Parameters;
+
+    beforeEach(() => {
+      parameters = parameters = {
         nf: {
           rating: [
             {
@@ -116,6 +148,36 @@ describe('parameter manager', () => {
               start: 4,
               end: 5,
               endInclusive: true,
+            },
+          ],
+        },
+        nfExcluded: {
+          price: [
+            {
+              state: 'excluded' as FacetValueState,
+              start: 10,
+              end: 20,
+              endInclusive: false,
+            },
+          ],
+        },
+        mnf: {
+          weight: [
+            {
+              state: 'selected' as FacetValueState,
+              start: 10,
+              end: 20,
+              endInclusive: false,
+            },
+          ],
+        },
+        mnfExcluded: {
+          height: [
+            {
+              state: 'excluded' as FacetValueState,
+              start: 10,
+              end: 20,
+              endInclusive: false,
             },
           ],
         },
@@ -131,16 +193,51 @@ describe('parameter manager', () => {
             },
           ],
         },
+        dfExcluded: {
+          expires: [
+            {
+              state: 'excluded' as FacetValueState,
+              start: '2010/01/01',
+              end: '2011/01/01',
+              endInclusive: false,
+            },
+          ],
+        },
         f: {
           size: ['small', 'medium'],
+        },
+        fExcluded: {
+          color: ['red'],
         },
         sortCriteria: buildRelevanceSortCriterion(),
         cf: {
           category: ['electronics'],
         },
       };
+    });
+    it('by default, #parameters is populated by the activeParametersSelector passed as a prop', () => {
+      mockActiveParametersSelector.mockReturnValue(parameters);
+
+      expect(parameterManager.state.parameters).toEqual(parameters);
+    });
+
+    it('when excludeDefaultParameters is set to false, #parameters is populated by the activeParametersSelector passed as a prop', () => {
+      initParameterManager({
+        excludeDefaultParameters: false,
+      });
 
       mockActiveParametersSelector.mockReturnValue(parameters);
+
+      expect(parameterManager.state.parameters).toEqual(parameters);
+    });
+
+    it('when excludeDefaultParameters is set to true, #parameters is populated by the commerceParameters slice', () => {
+      initEngine({...buildMockCommerceState(), commerceParameters: parameters});
+      initParameterManager({
+        excludeDefaultParameters: true,
+      });
+
+      mockActiveParametersSelector.mockReturnValue({});
 
       expect(parameterManager.state.parameters).toEqual(parameters);
     });