+++ packages/headless/src/features/generated-answer/generated-answer-analytics-actions.ts
@@ -70,19 +70,16 @@ export const logOpenGeneratedAnswerSource = (
     analyticsType: 'Rga.CitationClick',
     analyticsPayloadBuilder: (state): Rga.CitationClick | undefined => {
       const citation = citationSourceSelector(state, citationId);
-      const {id: rgaID} = generativeQuestionAnsweringIdSelector(state);
-      if (rgaID) {
-        return {
-          responseId: rgaID ?? '',
-          citationId,
-          itemMetadata: {
-            uniqueFieldName: 'permanentid',
-            uniqueFieldValue: citation?.permanentid ?? '',
-            title: citation?.title,
-            url: citation?.clickUri,
-          },
-        };
-      }
+      return {
+        responseId: state.search?.response.searchUid || '',
+        citationId,
+        itemMetadata: {
+          uniqueFieldName: 'permanentid',
+          uniqueFieldValue: citation?.permanentid ?? '',
+          title: citation?.title,
+          url: citation?.clickUri,
+        },
+      };
     },
   });
 
@@ -112,20 +109,17 @@ export const logHoverCitation = (
     analyticsType: 'Rga.CitationHover',
     analyticsPayloadBuilder: (state): Rga.CitationHover | undefined => {
       const citation = citationSourceSelector(state, citationId);
-      const {id: rgaID} = generativeQuestionAnsweringIdSelector(state);
-      if (rgaID) {
-        return {
-          responseId: rgaID,
-          citationId,
-          itemMetadata: {
-            uniqueFieldName: 'permanentid',
-            uniqueFieldValue: citation?.permanentid ?? '',
-            title: citation?.title,
-            url: citation?.clickUri,
-          },
-          citationHoverTimeInMs,
-        };
-      }
+      return {
+        responseId: state.search?.response.searchUid || '',
+        citationId,
+        itemMetadata: {
+          uniqueFieldName: 'permanentid',
+          uniqueFieldValue: citation?.permanentid ?? '',
+          title: citation?.title,
+          url: citation?.clickUri,
+        },
+        citationHoverTimeInMs,
+      };
     },
   });
 
@@ -146,13 +140,10 @@ export const logLikeGeneratedAnswer = (): CustomAction =>
     },
     analyticsType: 'Rga.AnswerAction',
     analyticsPayloadBuilder: (state): Rga.AnswerAction | undefined => {
-      const {id: rgaID} = generativeQuestionAnsweringIdSelector(state);
-      if (rgaID) {
-        return {
-          responseId: rgaID,
-          action: 'like',
-        };
-      }
+      return {
+        responseId: state.search?.response.searchUid || '',
+        action: 'like',
+      };
     },
   });
 
@@ -173,13 +164,10 @@ export const logDislikeGeneratedAnswer = (): CustomAction =>
     },
     analyticsType: 'Rga.AnswerAction',
     analyticsPayloadBuilder: (state): Rga.AnswerAction | undefined => {
-      const {id: rgaID} = generativeQuestionAnsweringIdSelector(state);
-      if (rgaID) {
-        return {
-          responseId: rgaID,
-          action: 'dislike',
-        };
-      }
+      return {
+        responseId: state.search?.response.searchUid || '',
+        action: 'dislike',
+      };
     },
   });
 
@@ -203,30 +191,27 @@ export const logGeneratedAnswerFeedback = (
     },
     analyticsType: 'Rga.SubmitFeedback',
     analyticsPayloadBuilder: (state): Rga.SubmitFeedback | undefined => {
-      const {id: rgaID} = generativeQuestionAnsweringIdSelector(state);
-      if (rgaID) {
-        const {
-          helpful,
-          readable,
-          documented,
-          hallucinationFree,
-          correctTopic,
-          documentUrl,
-          details,
-        } = feedback;
-        return {
-          responseId: rgaID,
-          helpful,
-          details: {
-            readable: parseEvaluationDetails(readable),
-            documented: parseEvaluationDetails(documented),
-            correctTopic: parseEvaluationDetails(correctTopic),
-            hallucinationFree: parseEvaluationDetails(hallucinationFree),
-          },
-          additionalNotes: details,
-          correctAnswerUrl: documentUrl,
-        };
-      }
+      const {
+        helpful,
+        readable,
+        documented,
+        hallucinationFree,
+        correctTopic,
+        documentUrl,
+        details,
+      } = feedback;
+      return {
+        responseId: state.search?.response.searchUid || '',
+        helpful,
+        details: {
+          readable: parseEvaluationDetails(readable),
+          documented: parseEvaluationDetails(documented),
+          correctTopic: parseEvaluationDetails(correctTopic),
+          hallucinationFree: parseEvaluationDetails(hallucinationFree),
+        },
+        additionalNotes: details,
+        correctAnswerUrl: documentUrl,
+      };
     },
   });
 
@@ -256,13 +241,10 @@ export const logGeneratedAnswerStreamEnd = (
     },
     analyticsType: 'Rga.StreamEnd',
     analyticsPayloadBuilder: (state): Rga.StreamEnd | undefined => {
-      const {id: rgaID} = generativeQuestionAnsweringIdSelector(state);
-      if (rgaID) {
-        return {
-          responseId: rgaID,
-          answerGenerated: true,
-        };
-      }
+      return {
+        responseId: state.search?.response.searchUid || '',
+        answerGenerated: true,
+      };
     },
   });
 
@@ -283,13 +265,10 @@ export const logGeneratedAnswerShowAnswers = (): CustomAction =>
     },
     analyticsType: 'Rga.AnswerAction',
     analyticsPayloadBuilder: (state): Rga.AnswerAction | undefined => {
-      const {id: rgaID} = generativeQuestionAnsweringIdSelector(state);
-      if (rgaID) {
-        return {
-          responseId: rgaID,
-          action: 'show',
-        };
-      }
+      return {
+        responseId: state.search?.response.searchUid || '',
+        action: 'show',
+      };
     },
   });
 
@@ -310,13 +289,10 @@ export const logGeneratedAnswerHideAnswers = (): CustomAction =>
     },
     analyticsType: 'Rga.AnswerAction',
     analyticsPayloadBuilder: (state): Rga.AnswerAction | undefined => {
-      const {id: rgaID} = generativeQuestionAnsweringIdSelector(state);
-      if (rgaID) {
-        return {
-          responseId: rgaID,
-          action: 'hide',
-        };
-      }
+      return {
+        responseId: state.search?.response.searchUid || '',
+        action: 'hide',
+      };
     },
   });
 
@@ -337,13 +313,10 @@ export const logGeneratedAnswerExpand = (): CustomAction =>
     },
     analyticsType: 'Rga.AnswerAction',
     analyticsPayloadBuilder: (state): Rga.AnswerAction | undefined => {
-      const {id: rgaID} = generativeQuestionAnsweringIdSelector(state);
-      if (rgaID) {
-        return {
-          responseId: rgaID,
-          action: 'expand',
-        };
-      }
+      return {
+        responseId: state.search?.response.searchUid || '',
+        action: 'expand',
+      };
     },
   });
 
@@ -364,13 +337,10 @@ export const logGeneratedAnswerCollapse = (): CustomAction =>
     },
     analyticsType: 'Rga.AnswerAction',
     analyticsPayloadBuilder: (state): Rga.AnswerAction | undefined => {
-      const {id: rgaID} = generativeQuestionAnsweringIdSelector(state);
-      if (rgaID) {
-        return {
-          responseId: rgaID,
-          action: 'collapse',
-        };
-      }
+      return {
+        responseId: state.search?.response.searchUid || '',
+        action: 'collapse',
+      };
     },
   });
 
@@ -391,13 +361,10 @@ export const logCopyGeneratedAnswer = (): CustomAction =>
     },
     analyticsType: 'Rga.AnswerAction',
     analyticsPayloadBuilder: (state): Rga.AnswerAction | undefined => {
-      const {id: rgaID} = generativeQuestionAnsweringIdSelector(state);
-      if (rgaID) {
-        return {
-          responseId: rgaID,
-          action: 'copyToClipboard',
-        };
-      }
+      return {
+        responseId: state.search?.response.searchUid || '',
+        action: 'copyToClipboard',
+      };
     },
   });