+++ packages/quantic/playwright.config.ts
@@ -28,26 +28,24 @@ export default defineConfig({
     : [['html'], ['list']],
   /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
   use: {
-    /* Base URL to use in actions like `await page.goto('/')`. */
-    baseURL: process.env.BASE_URL,
-
     /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
     trace: 'on-first-retry',
   },
 
-  /* Configure projects for major browsers */
   projects: [
     {
-      name: 'chromium',
-      use: {...devices['Desktop Chrome']},
+      name: 'LWS-enabled',
+      use: {
+        ...devices['Desktop Chrome'],
+        baseURL: process.env.Quantic__LWS_enabled_URL,
+      },
     },
-    // {
-    //   name: 'firefox',
-    //   use: {...devices['Desktop Firefox']},
-    // },
-    // {
-    //   name: 'webkit',
-    //   use: {...devices['Desktop Safari']},
-    // },
+    {
+      name: 'LWS-disabled',
+      use: {
+        ...devices['Desktop Chrome'],
+        baseURL: process.env.Quantic__LWS_disabled_URL,
+      },
+    },
   ],
 });