+++ packages/samples/headless-commerce-ssr-remix/app/routes/cart.purchase.tsx
@@ -1,0 +1,5 @@
+import externalCartService from '@/external-services/external-cart-service';
+
+export const action = async () => {
+  return await externalCartService.purchase();
+};